+++ hercules-ci-cli/src/Hercules/CLI/Client.hs
@@ -1,9 +1,40 @@
+{-# LANGUAGE BlockArguments #-}
 {-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DeriveAnyClass #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# OPTIONS_GHC -O0 #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
-module Hercules.CLI.Client where
+module Hercules.CLI.Client
+  ( -- * Setup
+    determineDefaultApiBaseUrl,
+    init,
+
+    -- * Using the client
+    HerculesClientToken (..),
+    HerculesClientEnv,
+    runHerculesClient,
+    runHerculesClient',
+    runHerculesClientEither,
+    runHerculesClientStream,
+
+    -- * Error handling
+    retryOnFail,
+    retryStreamOnFail,
+    shouldRetryClientError,
+    clientErrorSummary,
+    shouldRetryResponse,
+    waitRetryPolicy,
+    dieWithHttpError,
+    prettyPrintHttpErrors,
+
+    -- * Client function groups
+    accountsClient,
+    projectsClient,
+    reposClient,
+    stateClient,
+  )
+where
 
 -- TODO https://github.com/haskell-servant/servant/issues/986
 
@@ -15,8 +46,10 @@ import Hercules.API.Projects (ProjectsAPI)
 import Hercules.API.Repos (ReposAPI)
 import Hercules.API.State (ContentDisposition, ContentLength, RawBytes, StateAPI)
 import Hercules.Error
+import qualified Network.HTTP.Client as HTTP
 import qualified Network.HTTP.Client.TLS
-import Network.HTTP.Types.Status
+import Network.HTTP.Types.Status (Status (statusCode, statusMessage))
+import qualified Network.TLS as TLS
 import Protolude
 import RIO (RIO)
 import Servant.API
@@ -29,6 +62,8 @@ import Servant.Client.Generic (AsClientT)
 import Servant.Client.Streaming (ClientM, responseStatusCode, showBaseUrl)
 import qualified Servant.Client.Streaming
 import qualified System.Environment
+import qualified UnliftIO
+import UnliftIO.Retry (RetryPolicyM, RetryStatus, capDelay, fullJitterBackoff, retrying, rsIterNumber)
 
 -- | Bad instance to make it the client for State api compile. GHC seems to pick
 -- the wrong overlappable instance.
@@ -88,7 +123,7 @@ runHerculesClientStream ::
 runHerculesClientStream f g = do
   HerculesClientToken token <- asks getter
   HerculesClientEnv clientEnv <- asks getter
-  liftIO $ Servant.Client.Streaming.withClientM (f token) clientEnv g
+  liftIO $ convertInternalError $ Servant.Client.Streaming.withClientM (f token) clientEnv g
 
 runHerculesClient' :: (NFData a, Has HerculesClientEnv r) => Servant.Client.Streaming.ClientM a -> RIO r a
 runHerculesClient' = runHerculesClientEither' >=> escalate
@@ -96,7 +131,7 @@ runHerculesClient' = runHerculesClientEither' >=> escalate
 runHerculesClientEither' :: (NFData a, Has HerculesClientEnv r) => Servant.Client.Streaming.ClientM a -> RIO r (Either Servant.Client.Streaming.ClientError a)
 runHerculesClientEither' m = do
   HerculesClientEnv clientEnv <- asks getter
-  liftIO (Servant.Client.Streaming.runClientM m clientEnv)
+  liftIO $ convertInternalError $ Servant.Client.Streaming.runClientM m clientEnv
 
 init :: IO HerculesClientEnv
 init = do
@@ -152,6 +187,42 @@ shouldRetryClientError (ClientError.InvalidContentTypeHeader _) = False
 shouldRetryClientError (ClientError.ConnectionError _) = True
 shouldRetryClientError _ = False
 
+-- | A custom exception type for HTTP exceptions that we consider retryable.
+data HTTPInternalException = HTTPInternalException HTTP.Request SomeException
+  deriving (Exception, Show)
+
+-- | Convert a missed HTTP exception into a ClientError.
+-- This is useful for retrying.
+convertInternalError :: IO a -> IO a
+convertInternalError = deescalateInternalError >=> escalate
+
+deescalateInternalError :: IO a -> IO (Either ClientError a)
+deescalateInternalError m = handleJust matchClientException (pure . Left) (Right <$> m)
+  where
+    matchClientException =
+      fromException >=> \case
+        HTTP.HttpExceptionRequest req (HTTP.InternalException e)
+          | isJust (isRetryableHTTPInternalException e) ->
+              Just $ ClientError.ConnectionError $ toException $ HTTPInternalException req $ toException e
+        _ -> Nothing
+    isRetryableHTTPInternalException =
+      fromException >=> \e -> case e of
+        TLS.Terminated _mysteryBool _msg tlsError
+          | isRetryableTLSError tlsError ->
+              pass
+        _ -> Nothing
+    -- Error_Protocol has been observed. Most others are probably also worth retrying.
+    -- https://hackage.haskell.org/package/tls-1.9.0/docs/Network-TLS.html#t:TLSError
+    -- real world example: Error_Protocol ("remote side fatal error",True,BadRecordMac))
+    isRetryableTLSError =
+      \case
+        TLS.Error_Protocol {} -> True
+        TLS.Error_EOF {} -> True
+        TLS.Error_Packet {} -> True
+        TLS.Error_Packet_unexpected {} -> True
+        TLS.Error_Packet_Parsing {} -> True
+        _ -> False
+
 -- | ClientError printer that won't leak sensitive info.
 clientErrorSummary :: ClientError -> Text
 clientErrorSummary (ClientError.FailureResponse _ resp) = "status " <> show (responseStatusCode resp)
@@ -159,3 +230,48 @@ clientErrorSummary ClientError.DecodeFailure {} = "decode failure"
 clientErrorSummary ClientError.UnsupportedContentType {} = "unsupported content type"
 clientErrorSummary ClientError.InvalidContentTypeHeader {} = "invalid content type header"
 clientErrorSummary (ClientError.ConnectionError e) = "connection error: " <> show e
+
+simpleRetryPredicate :: (Applicative m) => (r -> Bool) -> RetryStatus -> r -> m Bool
+simpleRetryPredicate f _rs r = pure (f r)
+
+retryStreamOnFail ::
+  (Has HerculesClientToken r, Has HerculesClientEnv r) =>
+  Text ->
+  (Token -> ClientM b) ->
+  (Either ClientError b -> IO c) ->
+  RIO r c
+retryStreamOnFail shortDesc req handler = escalate =<< retryOnFailEither shortDesc (UnliftIO.try (runHerculesClientStream req handler))
+
+retryOnFail ::
+  (NFData b, Has HerculesClientToken r, Has HerculesClientEnv r) =>
+  Text ->
+  (Token -> ClientM b) ->
+  RIO r b
+retryOnFail shortDesc req = escalate =<< retryOnFailEither shortDesc (runHerculesClientEither req)
+
+retryOnFailEither ::
+  Text ->
+  RIO r (Either ClientError a) ->
+  RIO r (Either ClientError a)
+retryOnFailEither shortDesc req =
+  retrying
+    failureRetryPolicy
+    (simpleRetryPredicate shouldRetryResponse)
+    ( \rs -> do
+        when (rsIterNumber rs /= 0) do
+          liftIO $ putErrText $ "hci: " <> shortDesc <> " retrying."
+        r <- req
+        for_ (leftToMaybe r) \e -> do
+          liftIO $ putErrText $ "hci: " <> shortDesc <> " encountered " <> clientErrorSummary e <> "."
+          when (shouldRetryClientError e) do
+            liftIO $ putErrText $ "hci: " <> shortDesc <> " will retry."
+        pure r
+    )
+
+-- NB: fullJitterBackoff is not what it says it is: https://github.com/Soostone/retry/issues/46
+failureRetryPolicy :: (MonadIO m) => RetryPolicyM m
+failureRetryPolicy = capDelay (120 * 1000 * 1000) (fullJitterBackoff 100000)
+
+-- NB: fullJitterBackoff is not what it says it is: https://github.com/Soostone/retry/issues/46
+waitRetryPolicy :: (MonadIO m) => RetryPolicyM m
+waitRetryPolicy = capDelay (10 * 1000 * 1000) (fullJitterBackoff 500000)