+++ hercules-ci-cli/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+ - `hci state` Commands are now more robust, retrying when HTTP requests fail
+
 ## 0.3.6 - 2023-03-07
 
 ### Fixed